rename gateway_basic_trading.py to gateway_EVM_AMM.py
